Integrate MCP tools with OpenAI tool calling
