+++ app/db.py
@@ -1,0 +1,8 @@
+from .models import SessionLocal
+
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
