+++ requirements.txt
@@ -5,3 +5,4 @@ langgraph>=0.0.18
 google-auth-oauthlib>=1.1.0
 sqlalchemy>=2.0
 python-dotenv>=1.0.0
+pytest>=8.0.0
